+++ src/light_compressor/decompressors/__init__.py
@@ -1,22 +1,14 @@
-"""Simple stream readers for compressed buffers."""
+"""Simple stream decompressors."""
 
-from .decompressor import (
-    DecompressReader,
-    SnappyReader,
-)
 from .gzip import GZIPDecompressor
 from .lz4 import LZ4Decompressor
-from .reader import LimitedReader
 from .snappy import SNAPDecompressor
 from .zstd import ZSTDDecompressor
 
 
 __all__ = (
-    "DecompressReader",
     "GZIPDecompressor",
-    "LimitedReader",
     "LZ4Decompressor",
     "SNAPDecompressor",
-    "SnappyReader",
     "ZSTDDecompressor",
 )